pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hexlet-code"
version = "0.1.0"
description = "A utility to compare two configuration files and show the difference."
authors = [{ name = "Milena", email = "sr0.onlady@mail.ru" }]
dependencies = [
    "pytest",
    "flake8",
    "pytest-cov",
    "coverage",
    "pyyaml"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "flake8",
    "coverage"
]

[project.scripts]
gendiff = "gendiff.scripts.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

# Настройки для flake8
[tool.flake8]
max-line-length = 88
exclude = [".venv", "__pycache__", "build", "dist"]

# Настройки для pytest
#[tool.pytest.ini_options]
#addopts = "--cov=gendiff --cov-report=term-missing"
#testpaths = ["tests"]
#python_files = ["test_*.py"]

# Настройки для coverage
[tool.coverage.run]
branch = true
source = ["gendiff"]

[tool.coverage.report]
show_missing = true
skip_covered = true
